backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.1.0
pydantic==2.5.3
pikepdf==8.11.2
PyPDF2==3.0.1
aiofiles==23.2.1
